Give exegesis's rewrite object its own file
`exegesis.LspProxy.scala` held the proxy and an unrelated `object rewrite` of
sixteen inline hook constructors, an L2 violation; the latter would travel with
the proxy when it nests under `Lsp`. Verified on its own first.
